fix : refactoring into entity package
